+++ R/AllClasses.R
@@ -138,6 +138,7 @@ VariableCatalog <- setClass("VariableCatalog", contains="ShojiCatalog",
 DatasetCatalog <- setClass("DatasetCatalog", contains="ShojiCatalog")
 BatchCatalog <- setClass("BatchCatalog", contains="ShojiCatalog")
 PermissionCatalog <- setClass("PermissionCatalog", contains="ShojiCatalog")
+UserCatalog <- setClass("UserCatalog", contains="ShojiCatalog")
 
 default.useAlias <- function () {
     opt <- getOption("crunch.useAlias")