+++ tests/testthat/test-appending.R
@@ -93,7 +93,7 @@ if (run.integration.tests) {
                 })
             })
         })
-
+        
         try({
             file1 <- newDatasetFromFile(testfile.csv, name=now())
             file2 <- newDatasetFromFile(testfile.csv, name=now())
@@ -213,6 +213,14 @@ if (run.integration.tests) {
                 })
             })
         })
+
+                    # expect_message(try(appendDataset(part1, part2),
+                    #     silent=TRUE),
+                    #     paste("Critical conflict: Types do not match;", 
+                    #     "1 variable:", dQuote("v2")))
+                    # expect_error(appendDataset(part1, part2),
+                    #     "There are conflicts that cannot be resolved automatically.")
+                    # expect_identical(length(batches(part1)), 1L)
         
         datetime1 <- data.frame(
             cat=factor(c("A", "B")),